Add files via upload
First, I downloaded Exam and ran Apache and My SQL. 

Second, I set up the database on the phpMyadmin page. 

Third, I wrote PHP code to process data on the server and interact with the database, such as entering and retrieving data. 

Fourth, I wrote HTML code to create the page structure and display content such as buttons and titles. CSS to make the shape and colors more beautiful. 

Fifth, I finally opened this link in the browser and the interface appears like this.
 
You did an excellent job